fix(drone.cc):fix #5 implemented new update method
Now the update method will check to see which path the drone should take
when delivering the package. It uses the path attribute from json. Since
the umn.json doesn't have that attribute, I added it for now, will need
to discuss with the team on how to move forward