fix(customer.cc,package.cc):fix#31 update constructors
There was a change in the constructor that tried to access the radius
variable in the json object. Unforetunately, the tests did not have that
variable as the default radius worked for all our needs. simply added a
check in the constructors to ensure the attribute is there so it won't
seg fault.